Overhaul of the frontent's appearance.
Now has a background graphic to be displayed on all pages with all content blocks being semi-transparent to showcase the graphic.

Updated the FareShare logo for the navigation bar, the landing page and the signin/signup page to be an SVG with with the colour scheme.

Also added a small bugfix to change the map indicator pins to be state variables so that they immediately render rather than not appearing until after another interaction triggered an update after the initial page load.